nass-1359: make tabs
